pyproject.toml
[project]
name = "gami"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.61.0",
    "python-dotenv>=1.2.1",
]
